[admin] refactor(117) auth.type제거 후 상단에 type재정의, FSD 폴더 구조 변경
