+++ apps/admin/src/entity/auth/auth.key.ts
@@ -1,0 +1,3 @@
+export const authKeys = {
+  all: ['auth'] as const,
+}
