--- apps/admin/src/entity/auth/key.ts
@@ -1,4 +1,0 @@
-export const authKeys = {
-  all: ['auth'] as const,
-  profile: () => [...authKeys.all, 'profile'] as const,
-}
