--- apps/admin/src/entity/auth/types.ts
@@ -1,4 +1,0 @@
-export interface LoginResponse {
-  accessToken: string
-  refreshToken: string
-}
